fix(rename): missing rename
